requirements.txt
# Python >= 3.9
openai>=1.0.0,<2.0.0
httpx>=0.25.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
